in MpInputDeviceManager::disableDevice, when trying to acquire a device, break out of the mutex acquire loop if a device is not valid, or not in use. It is now more simplified, and description is provided for the #times through. 10ms wait time now used, as resolution of clocks generally isn't any higher than this.
Fix some compiler warnings (unsigned/signed, precision) in
MpInputDeviceManagerTest.



git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9198 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXmediaLib/src/test/mp/MpInputDeviceManagerTest.cpp
@@ -249,7 +249,7 @@ public:
         readerTask.start();
 
         // wait for frames to be read with slop time
-        int readerTaskWait = framePeriodMilliseconds * 1.5 * numBufferedFrames;
+        int readerTaskWait = (int)(framePeriodMilliseconds * 1.5 * numBufferedFrames);
         printf("waiting: %d for reader to finish\n", readerTaskWait);
         OsTask::delay(readerTaskWait);
         printf("done waiting for reader\n");
@@ -278,7 +278,7 @@ public:
 
             // Calculate the reference frame data
             actualSamples = storedSignal[frameIndex]->getSamples();
-            for(sampleIndex = 0; sampleIndex < samplesPerFrame; sampleIndex++)
+            for(sampleIndex = 0; (unsigned)sampleIndex < samplesPerFrame; sampleIndex++)
             {
                 referenceSamples[sampleIndex] =
                     MpSineWaveGeneratorDeviceDriver::calculateSample(frameTime,
